server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
stripe==7.11.0